수정,삭제 권한 표시 및 프레그먼트 뒤로가기

+++ app/src/main/java/com/acaroom/apicallpjt/fragment/board/BoardDetailFragment.kt
@@ -2,13 +2,17 @@ package com.acaroom.apicallpjt.fragment.board
 
 import android.animation.ObjectAnimator
 import android.content.Context
+import android.content.DialogInterface
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.GridLayoutManager
+import com.acaroom.apicallpjt.App
 import com.acaroom.apicallpjt.Config
 import com.acaroom.apicallpjt.MainActivity
 import com.acaroom.apicallpjt.R
@@ -40,7 +44,19 @@ class BoardDetailFragment : Fragment()  , MainActivity.onKeyBackPressedListener{
         savedInstanceState: Bundle?
     ): View? {
         val view = inflater.inflate(R.layout.fragment_board_detail, container, false)
+        var dialog = androidx.appcompat.app.AlertDialog.Builder(view.context)
+        dialog.setTitle("알림")
+
+
         val id = arguments?.getInt("id")!!
+        val userId = arguments?.getString("userId")!!
+
+        //자신이 쓴글이 아니면 수정,삭제 버튼 숨기기
+        if (App.prefs.userId != userId) {
+            view.fabMain.visibility = View.GONE
+            view.fabModify.visibility = View.GONE
+            view.fabRemove.visibility = View.GONE
+        }
 
 
         val recycleView = view.board_photo_list
@@ -54,11 +70,19 @@ class BoardDetailFragment : Fragment()  , MainActivity.onKeyBackPressedListener{
 //        recycleView.addItemDecoration(ver)
 //        recycleView.addItemDecoration(hor)
 
+
+
         view.fabMain.setOnClickListener{
             toggleFab()
         }
         view.fabRemove.setOnClickListener {
-            deleteBoard(id)
+            Log.i("클릭","!!")
+            dialog.setMessage("삭제 하시겠습니까?")
+            dialog.setPositiveButton("예", DialogInterface.OnClickListener() { _, _ ->
+                deleteBoard(id)
+            })
+            dialog.setNegativeButton("아니요" , DialogInterface.OnClickListener() {_,_->})
+            dialog.show()
         }
 
         boardService.findBoardDetail(id)
@@ -110,8 +134,8 @@ class BoardDetailFragment : Fragment()  , MainActivity.onKeyBackPressedListener{
                         val result = response.body()
                         if (result != null) {
                             Toast.makeText(context,"성공적으로 지웠음!" , Toast.LENGTH_SHORT).show()
-
                             var activity = (activity as MainActivity?)!!
+                            activity.supportFragmentManager.popBackStack("boardList" , FragmentManager.POP_BACK_STACK_INCLUSIVE)
                             activity.supportFragmentManager.beginTransaction()
                                 .replace(R.id.fragment, BoardFragment())
                                 .addToBackStack(null)
@@ -133,12 +157,12 @@ class BoardDetailFragment : Fragment()  , MainActivity.onKeyBackPressedListener{
 
         // 플로팅 액션 버튼 닫기 - 열려있는 플로팅 버튼 집어넣는 애니메이션 세팅
         if(isFabOpen) {
-            ObjectAnimator.ofFloat(fabWrite, "translationY", 0f).apply { start() }
+            ObjectAnimator.ofFloat(fabModify, "translationY", 0f).apply { start() }
             ObjectAnimator.ofFloat(fabRemove, "translationY", 0f).apply { start() }
             fabMain.setImageResource(R.drawable.ic_add)
         } else {
             // 플로팅 액션 버튼 열기 - 닫혀있는 플로팅 버튼 꺼내는 애니메이션 세팅
-            ObjectAnimator.ofFloat(fabWrite, "translationY", -200f).apply { start() }
+            ObjectAnimator.ofFloat(fabModify, "translationY", -200f).apply { start() }
             ObjectAnimator.ofFloat(fabRemove, "translationY", -400f).apply { start() }
             fabMain.setImageResource(R.drawable.ic_clear)
         }
